chore: moved checker to extensions
