--- tests/functional/c/consider/consider_refactoring_into_while_condition_py38.rc
@@ -1,2 +1,0 @@
-[testoptions]
-min_pyver=3.8
