+++ tests/functional/ext/consider_refactoring_into_while_condition/consider_refactoring_into_while_condition.py
