+++ tests/functional/ext/consider_refactoring_into_while_condition/consider_refactoring_into_while_condition_py38.rc
@@ -1,0 +1,4 @@
+[MAIN]
+load-plugins=pylint.extensions.consider_refactoring_into_while_condition,
+[testoptions]
+min_pyver=3.8
